refactor!: Rename projects to sync
