+++ docs/redirects.txt
@@ -15,3 +15,8 @@
 "parse/git.md" "url/git.md"
 "parse/hg.md" "url/hg.md"
 "parse/svn.md" "url/svn.md"
+"projects/index.md" "sync/index.md"
+"projects/base.md" "sync/base.md"
+"projects/git.md" "sync/git.md"
+"projects/hg.md" "sync/hg.md"
+"projects/svn.md" "sync/svn.md"
