+++ docs/sync/base.md
@@ -1,11 +1,11 @@
-# `libvcs.projects.base`
+# `libvcs.sync.base`
 
 Base objects / classes for projects.
 
 Adding your own VCS / Extending libvcs can be done through subclassing `BaseProject`.
 
 ```{eval-rst}
-.. automodule:: libvcs.projects.base
+.. automodule:: libvcs.sync.base
    :members:
    :show-inheritance:
 ```
