+++ docs/sync/hg.md
@@ -1,9 +1,9 @@
-# `libvcs.projects.hg`
+# `libvcs.sync.hg`
 
 For mercurial, aka `hg(1)`.
 
 ```{eval-rst}
-.. automodule:: libvcs.projects.hg
+.. automodule:: libvcs.sync.hg
    :members:
    :show-inheritance:
    :undoc-members:
