+++ docs/sync/svn.md
@@ -1,9 +1,9 @@
-# `libvcs.projects.svn`
+# `libvcs.sync.svn`
 
 For subversion, aka `svn(1)`
 
 ```{eval-rst}
-.. automodule:: libvcs.projects.svn
+.. automodule:: libvcs.sync.svn
    :members:
    :show-inheritance:
    :undoc-members:
