+++ src/libvcs/sync/constants.py
@@ -7,7 +7,7 @@ DEFAULT_VCS_LITERAL = Literal["git", "hg", "svn"]
 #: Union of VCS Classes
 DEFAULT_VCS_CLASS_UNION = type[Union[GitProject, MercurialProject, SubversionProject]]
 #: ``str`` -> ``class`` Map. ``DEFAULT_VCS_CLASS_MAP['git']`` ->
-#: :class:`~libvcs.projects.git.GitProject`
+#: :class:`~libvcs.sync.git.GitProject`
 DEFAULT_VCS_CLASS_MAP: dict[DEFAULT_VCS_LITERAL, DEFAULT_VCS_CLASS_UNION] = {
     "git": GitProject,
     "svn": SubversionProject,
