+++ tests/sync/test_hg.py
