+++ tests/url/test_svn.py
@@ -2,7 +2,7 @@ import typing
 
 import pytest
 
-from libvcs.projects.svn import SubversionProject
+from libvcs.sync.svn import SubversionProject
 from libvcs.url.base import MatcherRegistry
 from libvcs.url.svn import DEFAULT_MATCHERS, PIP_DEFAULT_MATCHERS, SvnURL
